Tratando o Erro 404

+++ src/main/java/br/com/alura/controller/TopicoController.java
@@ -2,6 +2,7 @@ package br.com.alura.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -60,25 +61,37 @@ public class TopicoController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalheTopicoDTO detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
+	public ResponseEntity<DetalheTopicoDTO> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalheTopicoDTO(topico.get()));
+		}
 		
-		return new DetalheTopicoDTO(topico);
+		return ResponseEntity.notFound().build();
 	}
 	
 	@Transactional
 	@PutMapping("/{id}")
 	public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-		Topico topico = form.atualizar(id, topicoRepository);
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDTO(topico));
+		}
 		
-		return ResponseEntity.ok(new TopicoDTO(topico));
+		return ResponseEntity.notFound().build();
 	}
 
 	@Transactional
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			
+			return ResponseEntity.ok().build();
+		}
 		
-		return ResponseEntity.ok().build();
+		return ResponseEntity.notFound().build();
 	}
 }
